Siirtele tiedostoja ja uudelleennimeä funktioita
