+++ backend/src/HyvaksymisEsitys/actions/tallenna.ts
@@ -1,12 +1,12 @@
 import * as API from "hassu-common/graphql/apiModel";
-import { auditLog } from "../logger";
-import { requirePermissionLuku, requirePermissionMuokkaa } from "../user";
+import { auditLog } from "../../logger";
+import { requirePermissionLuku, requirePermissionMuokkaa } from "../../user";
 import { IllegalArgumentError } from "hassu-common/error";
-import { adaptHyvaksymisEsitysToSave } from "./adaptHyvaksymisEsitysToSave";
-import { tallennaMuokattavaHyvaksymisEsitys } from "./dynamoDBCalls";
-import haeProjektinTiedotHyvaksymisEsityksesta, { HyvaksymisEsityksenTiedot } from "./dynamoDBCalls/get";
-import { AineistoNew } from "../database/model";
-import getHyvaksymisEsityksenAineistot from "./getAineistot";
+import { adaptHyvaksymisEsitysToSave } from "../adaptToSave/adaptHyvaksymisEsitysToSave";
+import { tallennaMuokattavaHyvaksymisEsitys } from "../dynamoDBCalls";
+import haeProjektinTiedotHyvaksymisEsityksesta, { HyvaksymisEsityksenTiedot } from "../dynamoDBCalls/getHyvaksymisEsityksenTiedot";
+import { AineistoNew } from "../../database/model";
+import getHyvaksymisEsityksenAineistot from "../getAineistot";
 
 export default async function tallennaHyvaksymisEsitys(input: API.TallennaHyvaksymisEsitysInput): Promise<string> {
   requirePermissionLuku();
