+++ backend/src/HyvaksymisEsitys/dynamoDBCalls/getHyvaksymisEsityksenTiedot.ts
